feat: added mobile navigation
