+++ src/components/Header.js
@@ -3,6 +3,8 @@ import { NavLink } from 'react-router-dom';
 import Media from 'react-media';
 import styled from 'styled-components';
 
+import Navigation from './Navigation';
+
 const SiteHeader = styled.header`
   display: flex;
   flex-direction: row;
@@ -10,18 +12,6 @@ const SiteHeader = styled.header`
   padding: 0 1rem;
   line-height: 1;
   background-color: #000;
-`;
-
-const SiteNavigation = styled.nav`
-  display: flex;
-  flex-grow: 1;
-
-  a,
-  a:hover,
-  a:focus {
-    line-height: 40px;
-    text-decoration: none;
-  }
 `;
 
 const SiteBrand = styled.div`
@@ -48,30 +38,6 @@ const SiteBrand = styled.div`
     font-size: 1rem;
   }
 `;
-
-const SiteNavigationLink = styled(NavLink)`
-  display: inline-block;
-  padding: 0 1rem;
-
-  &.router-link--active {
-    color: inherit;
-  }
-`;
-
-const Navigation = () => (
-  <SiteNavigation>
-    <SiteNavigationLink to="/videos" className="router-link" activeClassName="router-link--active">
-      {'this week\'s playlist'}
-    </SiteNavigationLink>
-    <SiteNavigationLink to="/howto" className="router-link" activeClassName="router-link--active">
-      add videos
-    </SiteNavigationLink>
-    {process.env.NODE_ENV !== 'production' &&
-      <SiteNavigationLink to="/playground" className="router-link" activeClassName="router-link--active">
-        component playground
-      </SiteNavigationLink>}
-  </SiteNavigation>
-);
 
 const Home = () => (
   <SiteHeader>
