+++ src/containers/App.js
@@ -1,9 +1,11 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
+import Media from 'react-media';
 import styled from 'styled-components';
 
 import Header from '../components/Header';
 import Footer from '../components/Footer';
+import MobileNav from '../components/MobileNav';
 
 const Wrapper = styled.div`
   height: 100%;
@@ -12,6 +14,7 @@ const Wrapper = styled.div`
 const AppContainer = ({ children }) => (
   <Wrapper>
     <Header />
+    <Media query="(max-width: 575px)" render={MobileNav} />
     {children}
     <Footer />
   </Wrapper>
